phase-3-chatbot/backend/requirements.txt
# FastAPI and Web Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlmodel==0.0.14
psycopg2-binary==2.9.9
alembic==1.13.0

# OpenAI and AI Agents
openai>=1.0.0

# Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Testing (add later)
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
# httpx>=0.25.0
